rename lib/password.js -> lib/bcrypt.js
The original name led to weird situations where it got confusing what
the variable was and what the library was.
