+++ lib/bcrypt.js
