+++ lib/passport.js
@@ -1,11 +1,11 @@
 import passport from 'passport';
 import LocalStrategy from 'passport-local';
 
+import bcrypt from './bcrypt';
 import {User} from '../models';
-import password from './password';
 import {AuthError} from './errors';
 
-const authenticate = function(username, passphrase, done) {
+function passwordAuth(username, password, done) {
 
   User
     .findOne({
@@ -13,12 +13,12 @@ const authenticate = function(username, passphrase, done) {
       attributes: {exclude: []},
     })
     .tap(user => { if (!user) throw new AuthError(); })
-    .tap(user => password.verify(passphrase, user.password))
+    .tap(user => bcrypt.verify(password, user.password))
     .then(user => done(null, user))
     .catch(AuthError, err => done(null, false, err))
     .catch(err => done(err));
-};
+}
 
-passport.use(new LocalStrategy(authenticate));
+passport.use(new LocalStrategy(passwordAuth));
 
 export default passport.authenticate('local', {session: false});
